Update types to fix type errors in NDLA frontend
