+++ packages/ndla-ui/src/Image/index.ts
@@ -9,5 +9,6 @@
 import Image from './Image';
 export { makeSrcQueryString } from './Image';
 export { ImageLink } from './ImageLink';
+export type { ImageCrop, ImageFocalPoint } from './Image';
 
 export default Image;
